requirements.txt
asgiref==3.2.7
autopep8==1.5.2
Django==3.0.6
django-cors-headers==3.2.1
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
pycodestyle==2.5.0
PyJWT==1.7.1
pytz==2020.1
sqlparse==0.3.1
gunicorn==19.7.1
whitenoise==3.3.1